Pulled the timezone enum tests out into a separate test file.
